Moved more stuff, can now parse all of common tests.
